+++ src/entity/Ball.java
@@ -1,6 +1,7 @@
 package entity;
 
-import core.Constants;
+import utils.Constants;
+import core.ResourceLoader;
 import utils.Vector2D;
 
 import java.awt.*;
@@ -13,6 +14,7 @@ public class Ball extends Entity{
         this.velocity = new Vector2D(1, -1).normalized().multiplied(Constants.BALL_SPEED);
         this.previousPosition = new Vector2D(x, y);
         this.radius = Constants.BALL_SIZE / 2;
+        this.img = ResourceLoader.loadImg("BallWarrior-master/assets/images/ball.png");
     }
 
     @Override
@@ -44,8 +46,12 @@ public class Ball extends Entity{
 
     @Override
     public void draw(Graphics g) {
-        g.setColor(Color.WHITE);
-        g.fillOval((int) position.x, (int) position.y, width, height);
+        if (img != null) {
+            g.drawImage(img,(int)position.x,(int)position.y,Constants.BALL_SIZE,Constants.BALL_SIZE,null);
+        }else {
+            g.setColor(Color.WHITE);
+            g.fillOval((int) position.x, (int) position.y, width, height);
+        }
     }
 
     public void bounceY() {