+++ src/entity/Paddle.java
@@ -1,17 +1,23 @@
 package entity;
 
-import core.Constants;
+import utils.Constants;
+import core.ResourceLoader;
 import core.InputHandler;
 
 import java.awt.*;
+import java.awt.image.BufferedImage;
 
 public class Paddle extends Entity {
 
-    private final InputHandler input;
+    private InputHandler input;
+    private BufferedImage[] sprites = new BufferedImage[2];
 
     public Paddle(float x, float y, InputHandler input) {
         super(x, y, Constants.PADDLE_WIDTH, Constants.PADDLE_HEIGHT);
         this.input = input;
+        sprites[0] = ResourceLoader.loadImg("BallWarrior-master/assets/images/paddle2.png");
+        sprites[1] = ResourceLoader.loadImg("BallWarrior-master/assets/images/paddle3.png");
+        this.setImg(sprites[0]);
     }
 
     @Override
@@ -39,7 +45,14 @@ public class Paddle extends Entity {
 
     @Override
     public void draw(Graphics g) {
-        g.setColor(Color.CYAN);
-        g.fillRect((int) position.x, (int) position.y, width, height);
+        if (img!=null) {
+            g.drawImage(img,(int)position.x,(int)position.y,Constants.PADDLE_WIDTH,Constants.PADDLE_HEIGHT,null);
+        }else {
+            g.setColor(Color.CYAN);
+            g.fillRect((int) position.x, (int) position.y, width, height);
+        }
+    }
+    public BufferedImage getSprite (int index) {
+        return sprites[index];
     }
 }