+++ src/game/collision/CircleVsAABB.java
@@ -1,9 +1,8 @@
 package game.collision;
 
-import core.Constants;
+import utils.Constants;
 import entity.Ball;
 import entity.Entity;
-import utils.MathUtils;
 import utils.Vector2D;
 
 /**